refactor: move funtions to shared_state.go
