+++ pkg/controller/azvolume.go
@@ -28,10 +28,7 @@ import (
 	"google.golang.org/grpc/status"
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
-	apiErrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	utilfeature "k8s.io/apiserver/pkg/util/feature"
-	"k8s.io/kubernetes/pkg/features"
 	azdiskv1beta2 "sigs.k8s.io/azuredisk-csi-driver/pkg/apis/azuredisk/v1beta2"
 	"sigs.k8s.io/azuredisk-csi-driver/pkg/azureutils"
 	util "sigs.k8s.io/azuredisk-csi-driver/pkg/util"
@@ -659,150 +656,6 @@ func NewAzVolumeController(mgr manager.Manager, volumeProvisioner VolumeProvisio
 	logger.V(2).Info("Controller set-up successful.")
 
 	return &reconciler, nil
-}
-
-func (c *SharedState) createAzVolumeFromPv(ctx context.Context, pv v1.PersistentVolume, annotations map[string]string) error {
-	var err error
-	ctx, w := workflow.New(ctx)
-	defer func() { w.Finish(err) }()
-
-	var desiredAzVolume *azdiskv1beta2.AzVolume
-	requiredBytes, _ := pv.Spec.Capacity.Storage().AsInt64()
-	volumeCapability := getVolumeCapabilityFromPv(&pv)
-
-	// create AzVolume CRI for CSI Volume Source
-	if pv.Spec.CSI != nil && pv.Spec.CSI.Driver == c.driverName {
-		desiredAzVolume, err = c.createAzVolumeFromCSISource(pv.Spec.CSI)
-		if err != nil {
-			return err
-		}
-		if azureutils.IsMultiNodePersistentVolume(pv) {
-			desiredAzVolume.Spec.MaxMountReplicaCount = 0
-		}
-
-		// create AzVolume CRI for AzureDisk Volume Source for migration case
-	} else if utilfeature.DefaultFeatureGate.Enabled(features.CSIMigration) &&
-		utilfeature.DefaultFeatureGate.Enabled(features.CSIMigrationAzureDisk) &&
-		pv.Spec.AzureDisk != nil {
-		desiredAzVolume = c.createAzVolumeFromAzureDiskVolumeSource(pv.Spec.AzureDisk)
-	}
-
-	if desiredAzVolume != nil {
-		if desiredAzVolume.Labels == nil {
-			desiredAzVolume.Labels = map[string]string{}
-		}
-		desiredAzVolume.Labels[consts.PvNameLabel] = pv.Name
-		if pv.Spec.ClaimRef != nil {
-			desiredAzVolume.Labels[consts.PvcNameLabel] = pv.Spec.ClaimRef.Name
-			desiredAzVolume.Labels[consts.PvcNamespaceLabel] = pv.Spec.ClaimRef.Namespace
-		}
-		desiredAzVolume.Spec.CapacityRange = &azdiskv1beta2.CapacityRange{RequiredBytes: requiredBytes}
-		desiredAzVolume.Spec.VolumeCapability = volumeCapability
-		desiredAzVolume.Spec.PersistentVolume = pv.Name
-		desiredAzVolume.Status.Annotations = annotations
-
-		w.AddDetailToLogger(consts.PvNameKey, pv.Name, consts.VolumeNameLabel, desiredAzVolume.Name)
-
-		w.Logger().Info("Creating AzVolume CRI")
-		if err = c.createAzVolume(ctx, desiredAzVolume); err != nil {
-			err = status.Errorf(codes.Internal, "failed to create AzVolume (%s) for PV (%s): %v", desiredAzVolume.Name, pv.Name, err)
-			return err
-		}
-	}
-	return nil
-}
-
-func (c *SharedState) createAzVolumeFromInline(ctx context.Context, inline *v1.AzureDiskVolumeSource) (err error) {
-	azVolume := c.createAzVolumeFromAzureDiskVolumeSource(inline)
-
-	if err = c.createAzVolume(ctx, azVolume); err != nil {
-		err = status.Errorf(codes.Internal, "failed to create AzVolume (%s) for inline (%s): %v", azVolume.Name, inline.DiskName, err)
-	}
-	return
-}
-
-func (c *SharedState) createAzVolumeFromCSISource(source *v1.CSIPersistentVolumeSource) (*azdiskv1beta2.AzVolume, error) {
-	diskName, err := azureutils.GetDiskName(source.VolumeHandle)
-	if err != nil {
-		return nil, fmt.Errorf("failed to extract diskName from volume handle (%s): %v", source.VolumeHandle, err)
-	}
-
-	_, maxMountReplicaCount := azureutils.GetMaxSharesAndMaxMountReplicaCount(source.VolumeAttributes, false)
-
-	var volumeParams map[string]string
-	if source.VolumeAttributes == nil {
-		volumeParams = make(map[string]string)
-	} else {
-		volumeParams = source.VolumeAttributes
-	}
-
-	azVolumeName := strings.ToLower(diskName)
-
-	azVolume := azdiskv1beta2.AzVolume{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:       azVolumeName,
-			Finalizers: []string{consts.AzVolumeFinalizer},
-		},
-		Spec: azdiskv1beta2.AzVolumeSpec{
-			MaxMountReplicaCount: maxMountReplicaCount,
-			Parameters:           volumeParams,
-			VolumeName:           diskName,
-		},
-		Status: azdiskv1beta2.AzVolumeStatus{
-			Detail: &azdiskv1beta2.AzVolumeStatusDetail{
-				VolumeID: source.VolumeHandle,
-			},
-			State: azdiskv1beta2.VolumeCreated,
-		},
-	}
-
-	return &azVolume, nil
-}
-
-func (c *SharedState) createAzVolumeFromAzureDiskVolumeSource(source *v1.AzureDiskVolumeSource) *azdiskv1beta2.AzVolume {
-	azVolume := azdiskv1beta2.AzVolume{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:       source.DiskName,
-			Finalizers: []string{consts.AzVolumeFinalizer},
-		},
-		Spec: azdiskv1beta2.AzVolumeSpec{
-			VolumeName:       source.DiskName,
-			VolumeCapability: []azdiskv1beta2.VolumeCapability{},
-		},
-		Status: azdiskv1beta2.AzVolumeStatus{
-			Detail: &azdiskv1beta2.AzVolumeStatusDetail{
-				VolumeID: source.DataDiskURI,
-			},
-			State:       azdiskv1beta2.VolumeCreated,
-			Annotations: map[string]string{consts.InlineVolumeAnnotation: source.DataDiskURI},
-		},
-	}
-
-	return &azVolume
-}
-
-func (c *SharedState) createAzVolume(ctx context.Context, desiredAzVolume *azdiskv1beta2.AzVolume) error {
-	var err error
-	var azVolume *azdiskv1beta2.AzVolume
-
-	if azVolume, err = c.azClient.DiskV1beta2().AzVolumes(c.objectNamespace).Get(ctx, desiredAzVolume.Name, metav1.GetOptions{}); err != nil {
-		if apiErrors.IsNotFound(err) {
-			if azVolume, err = c.azClient.DiskV1beta2().AzVolumes(c.objectNamespace).Create(ctx, desiredAzVolume, metav1.CreateOptions{}); err != nil {
-				return err
-			}
-		} else {
-			return err
-		}
-	}
-
-	updated := azVolume.DeepCopy()
-	updated.Status = desiredAzVolume.Status
-	if _, err := c.azClient.DiskV1beta2().AzVolumes(c.objectNamespace).UpdateStatus(ctx, updated, metav1.UpdateOptions{}); err != nil {
-		return err
-	}
-	// if AzVolume CRI successfully recreated, also recreate the operation queue for the volume
-	c.createOperationQueue(desiredAzVolume.Name)
-	return nil
 }
 
 func getVolumeCapabilityFromPv(pv *v1.PersistentVolume) []azdiskv1beta2.VolumeCapability {
